Refactor of project file structure
