+++ client/components/Forms/auth-form.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {connect} from 'react-redux'
 import PropTypes from 'prop-types'
-import {login} from '../store'
+import {login} from '../../store'
 
 /**
  * COMPONENT
