+++ client/components/Views/Cart.js
@@ -3,11 +3,11 @@ import React from 'react'
 import {connect} from 'react-redux'
 import {Link} from 'react-router-dom'
 import LocalCart from './LocalCart'
-import DeleteButton from './DeleteButton'
-import IncrementButton from './IncrementButton'
-import DecrementButton from './DecrementButton'
-import {findOrCreateCart} from '../store/cart'
-import SubmitOrder from './SubmitOrder'
+import DeleteButton from '../Buttons/DeleteButton'
+import IncrementButton from '../Buttons/IncrementButton'
+import DecrementButton from '../Buttons/DecrementButton'
+import {findOrCreateCart} from '../../store/redux/cart'
+import SubmitOrder from '../Buttons/SubmitOrder'
 
 class Cart extends React.Component {
   componentDidUpdate(prevProps, prevState) {
