+++ client/components/Views/user-home.js
@@ -2,7 +2,7 @@ import React, {useEffect} from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 import {Link} from 'react-router-dom'
-import {findOrCreateCart} from './../store/cart'
+import {findOrCreateCart} from '../../store/redux/cart'
 
 /**
  * COMPONENT
