+++ client/components/index.js
@@ -4,12 +4,12 @@
  * having to figure out which file they belong to!
  */
 export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
-export {Login} from './auth-form'
-export {Signup} from './Signup'
-export {default as AllProducts} from './AllProducts'
-export {default as AddToCart} from './AddToCart'
-export {default as SingleProduct} from './SingleProduct'
-export {default as UserList} from './UserList'
-export {default as Cart} from './Cart'
-export {default as OrderConfirmation} from './OrderConfirmation'
+export {default as UserHome} from './Views/user-home'
+export {Login} from './Forms/auth-form'
+export {Signup} from './Forms/Signup'
+export {default as AllProducts} from './Views/AllProducts'
+export {default as AddToCart} from './Buttons/AddToCart'
+export {default as SingleProduct} from './Views/SingleProduct'
+export {default as UserList} from './Views/UserList'
+export {default as Cart} from './Views/Cart'
+export {default as OrderConfirmation} from './Views/OrderConfirmation'
