+++ client/store/index.js
@@ -2,13 +2,13 @@ import {createStore, combineReducers, applyMiddleware} from 'redux'
 import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
-import user from './user'
-import allProductsReducer from './allProducts'
-import singleProductReducer from './singleProduct'
-import cartReducer from './cart'
-import allUsersReducer from './allUsers'
-import localCartReducer from './localCart'
-import {loadState, saveState} from './localStorage'
+import user from './redux/user'
+import allProductsReducer from './redux/allProducts'
+import singleProductReducer from './redux/singleProduct'
+import cartReducer from './redux/cart'
+import allUsersReducer from './redux/allUsers'
+import localCartReducer from './redux/localCart'
+import {loadState, saveState} from './redux/localStorage'
 
 const localCart = loadState()
 const reducer = combineReducers({
@@ -34,4 +34,4 @@ store.subscribe(() => {
 })
 
 export default store
-export * from './user'
+export * from './redux/user'
