+++ tests/UserList.spec.js
@@ -4,19 +4,21 @@ import {expect} from 'chai'
 import React from 'react'
 import enzyme, {shallow} from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import {UserHome} from './user-home'
+import UserList from '../client/components/Views/UserList'
 
 const adapter = new Adapter()
 enzyme.configure({adapter})
 
-describe('UserHome', () => {
-  let userHome
+describe('UserList', () => {
+  let userList
 
   beforeEach(() => {
-    userHome = shallow(<UserHome email="cody@email.com" />)
+    const wrapper = shallow(
+      <UserList firstName="Cody" lastName="Cody" email="cody@email.com" />
+    )
   })
 
   it('renders the email in an h3', () => {
-    expect(userHome.find('h3').text()).to.be.equal('Welcome, cody@email.com')
+    expect(wrapper.text()).to.be.include(' cody@email.com')
   })
 })
